添加简易Api接口

+++ database/migrations/2020_09_26_091420_create_categories_table.php
@@ -17,7 +17,7 @@ class CreateCategoriesTable extends Migration
             $table->bigIncrements('id');
             $table->string('name')->comment('名称');
             $table->text('description')->nullable()->comment('描述');
-            $table->integer('post_count')->default(0)->comment('帖子数');
+            $table->unsignedBigInteger('post_count')->default(0)->comment('帖子数');
             $table->timestamps();
         });
     }
