+++ database/migrations/2020_10_06_153700_create_replies_table.php
@@ -8,9 +8,9 @@ class CreateRepliesTable extends Migration
 	public function up()
 	{
 		Schema::create('replies', function(Blueprint $table) {
-            $table->increments('id');
-            $table->integer('topic_id')->unsigned()->default(0)->index();
-            $table->bigInteger('user_id')->unsigned()->default(0)->index();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('topic_id')->unsigned()->default(0)->index();
+            $table->unsignedBigInteger('user_id')->unsigned()->default(0)->index();
             $table->text('content')->nullable();
             $table->timestamps();
         });
